requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
confluent-kafka==2.5.0
httpx==0.27.0
kubernetes==29.1.0
pyyaml==6.0.1
psycopg2-binary==2.9.9
pydantic==2.8.2
